+++ sdk/resiliencemanagement/arm-resiliencemanagement/src/api/drillRuns/options.ts
@@ -1,17 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type {
-  DrillRunFailoverRequest,
-  DrillRunReprotectRequest,
-  ListReportDownloadUrlRequest,
-} from "../../models/models.js";
+import type { DrillRunFailoverRequest, DrillRunReprotectRequest } from "../../models/models.js";
 import type { OperationOptions } from "@azure-rest/core-client";
 
 /** Optional parameters. */
 export interface DrillRunsListReportDownloadUrlOptionalParams extends OperationOptions {
-  /** The content of the action request */
-  body?: ListReportDownloadUrlRequest;
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
 }
 
 /** Optional parameters. */