+++ sdk/resiliencemanagement/arm-resiliencemanagement/src/index.ts
@@ -204,7 +204,6 @@ export type {
   DrillRunAddNotesRequest,
   MarkAsCompleteRequest,
   ListReportDownloadUrlRequest,
-  ListReportDownloadUrlResponse,
   DrillRunResource,
   UnifiedResilienceItem,
   UnifiedResilienceItemProperties,